requirements.txt
atomicwrites==1.4.0
attrs==20.1.0
certifi==2020.6.20
colorama==0.4.3
cycler==0.10.0
iniconfig==1.0.1
kiwisolver==1.2.0
matplotlib==3.3.1
more-itertools==8.5.0
numpy==1.19.1
packaging==20.4
pandas==1.1.2
Pillow==7.2.0
pluggy==0.13.1
py==1.9.0
pyarrow==1.0.1
pyparsing==2.4.7
pytest==6.0.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
scipy==1.5.2
seaborn==0.11.0
six==1.15.0
toml==0.10.1
